correcting position of labels when matrix added
