+++ src/filter/NumberFilter.ts
@@ -45,8 +45,8 @@ export default class NumberFilter extends AVectorFilter<number, INumericalVector
   protected build(parent: HTMLElement) {
     const node = super.build(parent);
 
-    this.generateLabel(node);
-    this.generateTooltip(node);
+    this.generateLabel(node,this.data.desc.name);
+    this._toolTip = this.generateTooltip(node);
     this.generateDensityPlot(node);
 
 
@@ -70,37 +70,37 @@ export default class NumberFilter extends AVectorFilter<number, INumericalVector
   }
 
 
-  private generateLabel(node) {
-    const labelNode = d3.select(node).append('div').classed('filterlabel', true);
-    let name = this.data.desc.name;
-    if (name.length > 6) {
-      name = name.slice(0, 6) + '..';
-    }
-    const toolTip = (this.generateTooltip(node));
-    const fullName = this.data.desc.name;
-    labelNode.text(`${name.substring(0, 1).toUpperCase() + name.substring(1)}`)
-      .on('mouseover', function (d, i) {
-        toolTip.transition()
-          .duration(200)
-          .style('opacity', 1);
-        toolTip.html(`${fullName}`)
-          .style('left', ((<any>d3).event.pageX) + 'px')
-          .style('top', ((<any>d3).event.pageY - 10) + 'px');
-      })
-      .on('mouseout', function (d) {
-        toolTip.transition()
-          .duration(500)
-          .style('opacity', 0);
-      });
-  }
+  // private generateLabel(node) {
+  //   const labelNode = d3.select(node).append('div').classed('filterlabel', true);
+  //   let name = this.data.desc.name;
+  //   if (name.length > 6) {
+  //     name = name.slice(0, 6) + '..';
+  //   }
+  //   const toolTip = (this.generateTooltip(node));
+  //   const fullName = this.data.desc.name;
+  //   labelNode.text(`${name.substring(0, 1).toUpperCase() + name.substring(1)}`)
+  //     .on('mouseover', function (d, i) {
+  //       toolTip.transition()
+  //         .duration(200)
+  //         .style('opacity', 1);
+  //       toolTip.html(`${fullName}`)
+  //         .style('left', ((<any>d3).event.pageX) + 'px')
+  //         .style('top', ((<any>d3).event.pageY - 10) + 'px');
+  //     })
+  //     .on('mouseout', function (d) {
+  //       toolTip.transition()
+  //         .duration(500)
+  //         .style('opacity', 0);
+  //     });
+  // }
 
-  private generateTooltip(node) {
-    const tooltipDiv = d3.select(node).append('div')
-      .attr('class', 'tooltip')
-      .style('opacity', 0);
-    this._toolTip = tooltipDiv;
-    return tooltipDiv;
-  }
+  // private generateTooltip(node) {
+  //   const tooltipDiv = d3.select(node).append('div')
+  //     .attr('class', 'tooltip')
+  //     .style('opacity', 0);
+  //   this._toolTip = tooltipDiv;
+  //   return tooltipDiv;
+  // }
 
   private async getHistData() {
